refactor: add SnapTracker
