+++ packages/snap-account-service/src/SnapAccountService.test.ts
@@ -18,10 +18,6 @@ import type {
 } from './SnapAccountService';
 import { SnapAccountService } from './SnapAccountService';
 
-/**
- * The type of the messenger populated with all external actions and events
- * required by the service under test.
- */
 type RootMessenger = Messenger<
   MockAnyNamespace,
   MessengerActions<SnapAccountServiceMessenger>,
@@ -34,14 +30,10 @@ type MockKeyringControllerState = Pick<KeyringControllerState, 'keyrings'>;
 /** Mock truncated snap type for tests. */
 type MockTruncatedSnap = Pick<TruncatedSnap, 'id' | 'initialPermissions'>;
 
-/**
- * Mock objects for all external dependencies of {@link SnapAccountService}.
- */
 type Mocks = {
   // eslint-disable-next-line @typescript-eslint/naming-convention
   SnapController: {
     getState: jest.MockedFunction<() => SnapControllerState>;
-    getSnap: jest.MockedFunction<(snapId: string) => TruncatedSnap | null>;
     getRunnableSnaps: jest.MockedFunction<() => TruncatedSnap[]>;
   };
   // eslint-disable-next-line @typescript-eslint/naming-convention
@@ -144,84 +136,6 @@ function buildSnap(id: string, hasKeyring: boolean): TruncatedSnap {
 }
 
 /**
- * Publishes a `SnapController:snapInstalled` event on the root messenger.
- *
- * @param rootMessenger - The root messenger.
- * @param snap - The Snap that was installed.
- */
-function publishSnapInstalled(
-  rootMessenger: RootMessenger,
-  snap: TruncatedSnap,
-): void {
-  rootMessenger.publish('SnapController:snapInstalled', snap, 'origin', false);
-}
-
-/**
- * Publishes a `SnapController:snapEnabled` event on the root messenger.
- *
- * @param rootMessenger - The root messenger.
- * @param snap - The Snap that was enabled.
- */
-function publishSnapEnabled(
-  rootMessenger: RootMessenger,
-  snap: TruncatedSnap,
-): void {
-  rootMessenger.publish('SnapController:snapEnabled', snap);
-}
-
-/**
- * Publishes a `SnapController:snapDisabled` event on the root messenger.
- *
- * @param rootMessenger - The root messenger.
- * @param snap - The Snap that was disabled.
- */
-function publishSnapDisabled(
-  rootMessenger: RootMessenger,
-  snap: TruncatedSnap,
-): void {
-  rootMessenger.publish('SnapController:snapDisabled', snap);
-}
-
-/**
- * Publishes a `SnapController:snapBlocked` event on the root messenger.
- *
- * @param rootMessenger - The root messenger.
- * @param snapId - The ID of the Snap that was blocked.
- */
-function publishSnapBlocked(
-  rootMessenger: RootMessenger,
-  snapId: string,
-): void {
-  rootMessenger.publish('SnapController:snapBlocked', snapId);
-}
-
-/**
- * Publishes a `SnapController:snapUnblocked` event on the root messenger.
- *
- * @param rootMessenger - The root messenger.
- * @param snapId - The ID of the Snap that was unblocked.
- */
-function publishSnapUnblocked(
-  rootMessenger: RootMessenger,
-  snapId: string,
-): void {
-  rootMessenger.publish('SnapController:snapUnblocked', snapId);
-}
-
-/**
- * Publishes a `SnapController:snapUninstalled` event on the root messenger.
- *
- * @param rootMessenger - The root messenger.
- * @param snap - The Snap that was uninstalled.
- */
-function publishSnapUninstalled(
-  rootMessenger: RootMessenger,
-  snap: TruncatedSnap,
-): void {
-  rootMessenger.publish('SnapController:snapUninstalled', snap);
-}
-
-/**
  * Constructs the service under test with sensible defaults.
  *
  * @param args - The arguments to this function.
@@ -255,7 +169,6 @@ function setup({
       getState: jest
         .fn()
         .mockReturnValue({ isReady: snapIsReady } as SnapControllerState),
-      getSnap: jest.fn().mockReturnValue(null),
       getRunnableSnaps: jest.fn().mockReturnValue(runnableSnaps),
     },
     KeyringController: {
@@ -266,10 +179,6 @@ function setup({
   rootMessenger.registerActionHandler(
     'SnapController:getState',
     mocks.SnapController.getState,
-  );
-  rootMessenger.registerActionHandler(
-    'SnapController:getSnap',
-    mocks.SnapController.getSnap as never,
   );
   rootMessenger.registerActionHandler(
     'SnapController:getRunnableSnaps',
@@ -286,7 +195,6 @@ function setup({
 }
 
 const MOCK_SNAP_ID = 'npm:@metamask/mock-snap' as SnapId;
-const MOCK_OTHER_SNAP_ID = 'npm:@metamask/other-snap' as SnapId;
 
 describe('SnapAccountService', () => {
   describe('init', () => {
@@ -295,185 +203,17 @@ describe('SnapAccountService', () => {
 
       expect(await service.init()).toBeUndefined();
     });
-
-    it('seeds tracked Snaps from getRunnableSnaps, filtering out non-keyring Snaps', async () => {
-      const { service } = setup({
-        runnableSnaps: [
-          buildSnap(MOCK_SNAP_ID, true),
-          buildSnap(MOCK_OTHER_SNAP_ID, false),
-        ],
-      });
-
-      await service.init();
-
-      expect(service.getSnaps()).toStrictEqual([MOCK_SNAP_ID]);
-    });
   });
 
   describe('getSnaps', () => {
-    it('returns an empty array before init', () => {
+    it('exposes tracked Snaps seeded by init', async () => {
       const { service } = setup({
         runnableSnaps: [buildSnap(MOCK_SNAP_ID, true)],
       });
 
-      expect(service.getSnaps()).toStrictEqual([]);
-    });
-  });
-
-  describe('lifecycle events', () => {
-    it('ignores add events received before init', async () => {
-      const { service, rootMessenger } = setup();
-
-      publishSnapInstalled(rootMessenger, buildSnap(MOCK_SNAP_ID, true));
-
-      await service.init();
-
-      expect(service.getSnaps()).toStrictEqual([]);
-    });
-
-    it('ignores remove events received before init', async () => {
-      const { service, rootMessenger } = setup({
-        runnableSnaps: [buildSnap(MOCK_SNAP_ID, true)],
-      });
-
-      // Publish a removal event *before* init: it must be ignored, so once
-      // init seeds from `getRunnableSnaps` the Snap is still tracked.
-      publishSnapUninstalled(rootMessenger, buildSnap(MOCK_SNAP_ID, true));
-
       await service.init();
 
       expect(service.getSnaps()).toStrictEqual([MOCK_SNAP_ID]);
-    });
-
-    it('adds a Snap on snapInstalled when it has the keyring endowment', async () => {
-      const { service, rootMessenger } = setup();
-
-      await service.init();
-      publishSnapInstalled(rootMessenger, buildSnap(MOCK_SNAP_ID, true));
-
-      expect(service.getSnaps()).toStrictEqual([MOCK_SNAP_ID]);
-    });
-
-    it('does not add a Snap on snapInstalled when it lacks the keyring endowment', async () => {
-      const { service, rootMessenger } = setup();
-
-      await service.init();
-      publishSnapInstalled(rootMessenger, buildSnap(MOCK_SNAP_ID, false));
-
-      expect(service.getSnaps()).toStrictEqual([]);
-    });
-
-    it('adds a Snap on snapEnabled when it has the keyring endowment', async () => {
-      const { service, rootMessenger } = setup();
-
-      await service.init();
-      publishSnapEnabled(rootMessenger, buildSnap(MOCK_SNAP_ID, true));
-
-      expect(service.getSnaps()).toStrictEqual([MOCK_SNAP_ID]);
-    });
-
-    it('removes a Snap on snapDisabled', async () => {
-      const { service, rootMessenger } = setup({
-        runnableSnaps: [buildSnap(MOCK_SNAP_ID, true)],
-      });
-
-      await service.init();
-      expect(service.getSnaps()).toStrictEqual([MOCK_SNAP_ID]);
-
-      publishSnapDisabled(rootMessenger, buildSnap(MOCK_SNAP_ID, true));
-
-      expect(service.getSnaps()).toStrictEqual([]);
-    });
-
-    it('removes a Snap on snapBlocked', async () => {
-      const { service, rootMessenger } = setup({
-        runnableSnaps: [buildSnap(MOCK_SNAP_ID, true)],
-      });
-
-      await service.init();
-      publishSnapBlocked(rootMessenger, MOCK_SNAP_ID);
-
-      expect(service.getSnaps()).toStrictEqual([]);
-    });
-
-    it('ignores snapUnblocked received before init', async () => {
-      const { service, rootMessenger, mocks } = setup();
-
-      mocks.SnapController.getSnap.mockReturnValue({
-        ...buildSnap(MOCK_SNAP_ID, true),
-        enabled: true,
-        blocked: false,
-      } as TruncatedSnap);
-      publishSnapUnblocked(rootMessenger, MOCK_SNAP_ID);
-
-      await service.init();
-
-      expect(service.getSnaps()).toStrictEqual([]);
-    });
-
-    it('re-adds a Snap on snapUnblocked when it is enabled and has the keyring endowment', async () => {
-      const { service, rootMessenger, mocks } = setup();
-
-      await service.init();
-      mocks.SnapController.getSnap.mockReturnValue({
-        ...buildSnap(MOCK_SNAP_ID, true),
-        enabled: true,
-        blocked: false,
-      } as TruncatedSnap);
-
-      publishSnapUnblocked(rootMessenger, MOCK_SNAP_ID);
-
-      expect(service.getSnaps()).toStrictEqual([MOCK_SNAP_ID]);
-    });
-
-    it('does not re-add a Snap on snapUnblocked when it is disabled', async () => {
-      const { service, rootMessenger, mocks } = setup();
-
-      await service.init();
-      mocks.SnapController.getSnap.mockReturnValue({
-        ...buildSnap(MOCK_SNAP_ID, true),
-        enabled: false,
-        blocked: false,
-      } as TruncatedSnap);
-
-      publishSnapUnblocked(rootMessenger, MOCK_SNAP_ID);
-
-      expect(service.getSnaps()).toStrictEqual([]);
-    });
-
-    it('does not re-add a Snap on snapUnblocked when it lacks the keyring endowment', async () => {
-      const { service, rootMessenger, mocks } = setup();
-
-      await service.init();
-      mocks.SnapController.getSnap.mockReturnValue({
-        ...buildSnap(MOCK_SNAP_ID, false),
-        enabled: true,
-        blocked: false,
-      } as TruncatedSnap);
-
-      publishSnapUnblocked(rootMessenger, MOCK_SNAP_ID);
-
-      expect(service.getSnaps()).toStrictEqual([]);
-    });
-
-    it('does not re-add a Snap on snapUnblocked when getSnap returns null', async () => {
-      const { service, rootMessenger } = setup();
-
-      await service.init();
-      publishSnapUnblocked(rootMessenger, MOCK_SNAP_ID);
-
-      expect(service.getSnaps()).toStrictEqual([]);
-    });
-
-    it('removes a Snap on snapUninstalled', async () => {
-      const { service, rootMessenger } = setup({
-        runnableSnaps: [buildSnap(MOCK_SNAP_ID, true)],
-      });
-
-      await service.init();
-      publishSnapUninstalled(rootMessenger, buildSnap(MOCK_SNAP_ID, true));
-
-      expect(service.getSnaps()).toStrictEqual([]);
     });
   });
