+++ packages/snap-account-service/src/SnapAccountService.ts
@@ -16,26 +16,14 @@ import type {
   SnapControllerStateChangeEvent,
 } from '@metamask/snaps-controllers';
 import { SnapId } from '@metamask/snaps-sdk';
-import type { TruncatedSnap } from '@metamask/snaps-utils';
 
-import { projectLogger as log } from './logger';
 import type {
   SnapAccountServiceEnsureReadyAction,
   SnapAccountServiceGetSnapsAction,
 } from './SnapAccountService-method-action-types';
 import { SnapPlatformWatcher } from './SnapPlatformWatcher';
 import type { SnapPlatformWatcherConfig } from './SnapPlatformWatcher';
-
-/**
- * Checks if a given Snap is an account management Snap.
- *
- * @param snap - The Snap to check.
- * @returns True if the Snap declares the `endowment:keyring` initial
- * permission.
- */
-function isAccountManagementSnap(snap: TruncatedSnap): boolean {
-  return snap.initialPermissions['endowment:keyring'] !== undefined;
-}
+import { SnapTracker } from './SnapTracker';
 
 /**
  * The name of the {@link SnapAccountService}, used to namespace the service's
@@ -121,9 +109,7 @@ export class SnapAccountService {
 
   readonly #watcher: SnapPlatformWatcher;
 
-  readonly #snaps: Set<SnapId> = new Set();
-
-  #initialized = false;
+  readonly #tracker: SnapTracker;
 
   /**
    * Constructs a new {@link SnapAccountService}.
@@ -139,25 +125,7 @@ export class SnapAccountService {
       messenger,
       config?.snapPlatformWatcher,
     );
-
-    this.#messenger.subscribe('SnapController:snapInstalled', (snap) =>
-      this.#handleSnapAdded(snap, 'installed'),
-    );
-    this.#messenger.subscribe('SnapController:snapUninstalled', (snap) =>
-      this.#handleSnapRemoved(snap.id, 'uninstalled'),
-    );
-    this.#messenger.subscribe('SnapController:snapEnabled', (snap) =>
-      this.#handleSnapAdded(snap, 'enabled'),
-    );
-    this.#messenger.subscribe('SnapController:snapDisabled', (snap) =>
-      this.#handleSnapRemoved(snap.id, 'disabled'),
-    );
-    this.#messenger.subscribe('SnapController:snapBlocked', (snapId) =>
-      this.#handleSnapRemoved(snapId as SnapId, 'blocked'),
-    );
-    this.#messenger.subscribe('SnapController:snapUnblocked', (snapId) =>
-      this.#handleSnapUnblocked(snapId as SnapId),
-    );
+    this.#tracker = new SnapTracker(messenger);
 
     this.#messenger.registerMethodActionHandlers(
       this,
@@ -173,15 +141,7 @@ export class SnapAccountService {
    * events.
    */
   async init(): Promise<void> {
-    const runnable = this.#messenger.call('SnapController:getRunnableSnaps');
-    for (const snap of runnable) {
-      if (isAccountManagementSnap(snap) && !this.#snaps.has(snap.id)) {
-        log(`Found account management Snap: ${snap.id} (initialization)`);
-        this.#snaps.add(snap.id);
-      }
-    }
-
-    this.#initialized = true;
+    await this.#tracker.init();
   }
 
   /**
@@ -192,7 +152,7 @@ export class SnapAccountService {
    * @returns The IDs of tracked account-management Snaps.
    */
   getSnaps(): SnapId[] {
-    return [...this.#snaps];
+    return this.#tracker.getSnaps();
   }
 
   /**
@@ -206,66 +166,11 @@ export class SnapAccountService {
    * @throws If `snapId` is not a tracked account-management Snap.
    */
   async ensureReady(snapId: SnapId): Promise<void> {
-    if (!this.#snaps.has(snapId)) {
+    if (!this.#tracker.has(snapId)) {
       throw new Error(`Unknown snap: "${snapId}"`);
     }
     // Before doing anything with our Snap, we need to make sure the platform
     // is ready to process requests.
     await this.#watcher.ensureCanUseSnapPlatform();
-  }
-
-  /**
-   * Handles a Snap being added (installed or enabled). If the Snap is an
-   * account-management Snap, adds it to the internal set of tracked Snaps.
-   *
-   * @param snap - The Snap that was installed or enabled.
-   * @param reason - The reason the Snap was added.
-   */
-  #handleSnapAdded(snap: TruncatedSnap, reason: string): void {
-    if (!this.#initialized) {
-      return;
-    }
-
-    if (isAccountManagementSnap(snap) && !this.#snaps.has(snap.id)) {
-      log(`Added account management Snap: ${snap.id} (${reason})`);
-
-      this.#snaps.add(snap.id);
-    }
-  }
-
-  /**
-   * Handles a Snap being unblocked. If the Snap is an enabled
-   * account-management Snap, re-adds it to the internal set of tracked Snaps.
-   *
-   * @param snapId - The Snap ID that was unblocked.
-   */
-  #handleSnapUnblocked(snapId: SnapId): void {
-    if (!this.#initialized) {
-      return;
-    }
-
-    const snap = this.#messenger.call('SnapController:getSnap', snapId);
-    if (snap && snap.enabled && !snap.blocked) {
-      this.#handleSnapAdded(snap, 'unblocked');
-    }
-  }
-
-  /**
-   * Handles a Snap being removed (disabled, blocked, or uninstalled). If the Snap is an
-   * account-management Snap, removes it from the internal set of tracked Snaps.
-   *
-   * @param snapId - The Snap ID that was disabled, blocked, or uninstalled.
-   * @param reason - The reason the Snap was removed.
-   */
-  #handleSnapRemoved(snapId: SnapId, reason: string): void {
-    if (!this.#initialized) {
-      return;
-    }
-
-    if (this.#snaps.has(snapId)) {
-      log(`Removed account management Snap: ${snapId} (${reason})`);
-
-      this.#snaps.delete(snapId);
-    }
   }
 }
